# flowinfra: use the flow context in FlowStream RPC in outbox