--- a/pkg/sql/flowinfra/outbox_test.go
+++ b/pkg/sql/flowinfra/outbox_test.go
@@ -122,7 +122,7 @@ func TestOutbox(t *testing.T) {
 			// Now send another row that the outbox will discard.
 			row = rowenc.EncDatumRow{rowenc.DatumToEncDatum(types.Int, tree.NewDInt(tree.DInt(2)))}
 			if consumerStatus := outbox.Push(row, nil /* meta */); consumerStatus != execinfra.DrainRequested {
-				return errors.Errorf("expected status: %d, got: %d", execinfra.NeedMoreRows, consumerStatus)
+				return errors.Errorf("expected status: %d, got: %d", execinfra.DrainRequested, consumerStatus)
 			}
 
 			// Send some metadata.
@@ -206,10 +206,10 @@ func TestOutbox(t *testing.T) {
 		t.Errorf("invalid results: %s, expected %s'", str, expected)
 	}
 
-	// The outbox should shut down since the producer closed.
-	outboxWG.Wait()
 	// Signal the server to shut down the stream.
 	streamNotification.Donec <- nil
+	// The outbox should shut down since the stream is closed.
+	outboxWG.Wait()
 }
 
 // Test that an outbox connects its stream as soon as possible (i.e. before
@@ -559,8 +559,8 @@ func BenchmarkOutbox(b *testing.B) {
 				}
 			}
 			outbox.ProducerDone()
-			outboxWG.Wait()
 			streamNotification.Donec <- nil
+			outboxWG.Wait()
 		})
 	}
 }