--- a/pkg/sql/flowinfra/utils_test.go
+++ b/pkg/sql/flowinfra/utils_test.go
@@ -12,6 +12,8 @@ package flowinfra
 
 import (
 	"context"
+	"io"
+	"testing"
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
@@ -23,44 +25,58 @@ import (
 )
 
 // createDummyStream creates the server and client side of a FlowStream stream.
-// This can be use by tests to pretend that then have received a FlowStream RPC.
+// This can be use by tests to pretend that they have received a FlowStream RPC.
 // The stream can be used to send messages (ConsumerSignal's) on it (within a
 // gRPC window limit since nobody's reading from the stream), for example
 // Handshake messages.
 //
 // We do this by creating a mock server, dialing into it and capturing the
 // server stream. The server-side RPC call will be blocked until the caller
-// calls the returned cleanup function.
-func createDummyStream() (
+// calls the returned cleanup function. The cleanup function also "drains" the
+// client-side stream.
+func createDummyStream(
+	t *testing.T,
+) (
 	serverStream execinfrapb.DistSQL_FlowStreamServer,
 	clientStream execinfrapb.DistSQL_FlowStreamClient,
 	cleanup func(),
-	err error,
 ) {
 	stopper := stop.NewStopper()
 	ctx := context.Background()
 	clock := hlc.NewClockWithSystemTimeSource(time.Nanosecond /* maxOffset */)
 	storageClusterID, mockServer, addr, err := execinfrapb.StartMockDistSQLServer(ctx, clock, stopper, execinfra.StaticSQLInstanceID)
 	if err != nil {
-		return nil, nil, nil, err
+		t.Fatal(err)
 	}
 
 	rpcContext := rpc.NewInsecureTestingContextWithClusterID(ctx, clock, stopper, storageClusterID)
 	conn, err := rpcContext.GRPCDialNode(addr.String(), roachpb.NodeID(execinfra.StaticSQLInstanceID),
 		rpc.DefaultClass).Connect(ctx)
 	if err != nil {
-		return nil, nil, nil, err
+		t.Fatal(err)
 	}
 	client := execinfrapb.NewDistSQLClient(conn)
 	clientStream, err = client.FlowStream(ctx)
 	if err != nil {
-		return nil, nil, nil, err
+		t.Fatal(err)
 	}
 	streamNotification := <-mockServer.InboundStreams
 	serverStream = streamNotification.Stream
 	cleanup = func() {
 		close(streamNotification.Donec)
+		// After the RPC is unblocked, we have to drain the client side in order
+		// to simulate what happens in production (where the watchdog goroutine
+		// of the outbox does this).
+		for {
+			_, err := clientStream.Recv()
+			if err != nil {
+				if err == io.EOF {
+					break
+				}
+				t.Fatal(err)
+			}
+		}
 		stopper.Stop(ctx)
 	}
-	return serverStream, clientStream, cleanup, nil
+	return serverStream, clientStream, cleanup
 }