Previously, we used the background context when performing the
`FlowStream` RPC in the row-based outbox rather than the flow context.
This has been the case for like four years now, but I believe it is
incorrect. In particular, if the flow context is canceled, this would
not shutdown the gRPC stream, possibly leaking some of the goroutines
(e.g. the goroutine listening for the drain signal from the inbox). This
is now fixed by correctly using the flow context.

This commit also fixes the outbox to satisfy the gRPC contract hidden in
[1] where the listener goroutine would not `Recv` everything from the
stream which could result in leakage of gRPC goroutines.

This required the cleanup of the outbox infrastructure to bring it more
in line with what we have in the vectorized engine (which has been
updated some time ago with recent learnings). In particular, the
following problems have been fixed:
- `CloseSend` signal from the main goroutine of the outbox previously
could have been omitted in some cases (I'm not sure what the impact of
this is though)
- the listener ("watchdog") goroutine is now tracked by the wait group
of the flow which makes it so that `Flow.Wait` blocks until the watchdog
exits. That goroutine now exits only after it `Recv`es an error from the
server side (the inbox) of `FlowStream` RPC.

[1] https://pkg.go.dev/google.golang.org/grpc#ClientConn.NewStream

Informs: #90841.

Epic: None

Release note: None