@yuzefovich 2022-12-20
I think the CI should be green with the latest push (there was a hang in a benchmark previously) plus I added another commit to fix a test (exposed in #93840). RFAL.

Not that it matters much, but the benchmarks seem to have improved:
```
name                 old time/op    new time/op     delta
Outbox/numCols=1-24     205ns ±32%      216ns ±15%     ~     (p=0.511 n=10+9)
Outbox/numCols=2-24     271ns ± 4%      224ns ± 3%  -17.10%  (p=0.000 n=10+10)
Outbox/numCols=4-24     331ns ± 4%      280ns ± 2%  -15.29%  (p=0.000 n=10+10)
Outbox/numCols=8-24     464ns ± 3%      406ns ± 2%  -12.53%  (p=0.000 n=10+10)

name                 old speed      new speed       delta
Outbox/numCols=1-24  39.9MB/s ±43%   37.3MB/s ±17%     ~     (p=0.497 n=10+9)
Outbox/numCols=2-24  59.1MB/s ± 4%   71.3MB/s ± 3%  +20.62%  (p=0.000 n=10+10)
Outbox/numCols=4-24  96.9MB/s ± 4%  114.3MB/s ± 2%  +17.99%  (p=0.000 n=10+10)
Outbox/numCols=8-24   138MB/s ± 3%    158MB/s ± 2%  +14.30%  (p=0.000 n=10+10)

name                 old alloc/op   new alloc/op    delta
Outbox/numCols=1-24      275B ±54%       274B ±54%     ~     (p=0.811 n=10+10)
Outbox/numCols=2-24      379B ± 4%       433B ± 2%  +14.10%  (p=0.000 n=10+9)
Outbox/numCols=4-24      462B ± 5%       544B ± 1%  +17.89%  (p=0.000 n=10+10)
Outbox/numCols=8-24      646B ± 4%       773B ± 2%  +19.61%  (p=0.000 n=10+9)

name                 old allocs/op  new allocs/op   delta
Outbox/numCols=1-24      3.56 ±44%       3.56 ±44%     ~     (p=1.000 n=9+9)
Outbox/numCols=2-24      5.00 ± 0%       6.00 ± 0%  +20.00%  (p=0.000 n=10+9)
Outbox/numCols=4-24      6.00 ± 0%       7.00 ± 0%  +16.67%  (p=0.000 n=10+10)
Outbox/numCols=8-24      8.70 ± 8%      10.70 ± 7%  +22.99%  (p=0.000 n=10+10)
```