@yuzefovich 2022-12-21
Alright, updated the commit message. I don't think another look is needed, so will merge on green CI.